WIP: adding integration testing to the hub web project.
