+++ research-hub-web/package.json
@@ -21,7 +21,7 @@
     "generate-watch": "graphql-codegen --watch './src/app/graphql/queries/*.graphql'",
     "graphql-server": "nodemon ../cer-graphql/index.js",
     "graphql-server-watch": "npm run dev --prefix ../cer-graphql",
-    "integration-tests": "./node_modules/.bin/protractor protractor.conf.js --specs e2e/*.integration-spec.ts"
+    "integration-tests": "./node_modules/.bin/protractor protractor.conf.js --specs e2e/*.integration-spec.*"
   },
   "private": true,
   "dependencies": {
